Arrays!!
Class Examples with arrays and Homework with arrays

A simple Income calculator that determines who has the largest income
and who was the lowest income.

A simple grading program for a teacher.  The teacher enters the number
of students and grades then calculates the best grade and what the
letter grades are based on a formula.